requirements.txt
aiohttp==3.6.2
alembic==1.4.2
asyncpgsa==0.26.3  # Эта либа сама установит правильные asyncpg и sqlalchemy
pytest==6.0.1
python-dotenv==0.14.0
validated-dc==1.2.1
